Adds an ApiConnectionManager for contacting our SMS API
Summary:
We want a nice, uniform way to interact with Faraday and the parameters
we think we need to send over HTTP. This service also provides us with
some helpful trouble shooting messages in case we mis-configure our
application.

I suspect we would extend the parameter attributes in this class if we
start expanding what our SMS API accepts.