Adds data model for verifying phones
Summary:

We want to be able to store phone numbers and their relevant
verification codes.  These items can expire.  A phone may need to
re-verify if it expires. A new code may need to get issued if it
expires.

Currently, we only support US and Canadian numbers.  You can extend the
supported countries by modifying the constant in Phone.