Adds SmsPhoneVerification for handling the phone verification workflow
Summary:

The SmsPhoneVerification class allows us to capture a single
responsibility for creating the state required to perform a phone
verification.  I would have preferred to call the model
PhoneVerification since a code may not necessarily come from Sms since
we can inject our own API instance that could say do a robo-call or
something else.  PhoneVerification appears to be the name the project
was generated with, so it is reserved.